requirements.txt
django==2.2.16
django-cors-headers==3.13.0
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
requests==2.26.0
Pillow==8.3.1
sqlparse==0.3.1
sorl-thumbnail==12.7.0
